Use CC instead of LD for OBJCOPY
When cross-compiling for mips targets using Buildroot, using LD for
partial linking fails in the following way:
/tmp/instance-1/output-1/host/opt/ext-toolchain/bin/../lib/gcc/mips64el-buildroot-linux-uclibc/8.4.0/../../../../mips64el-buildroot-linux-uclibc/bin/ld: build/release/libmupdf.a(Dingbats.cff.o): ABI is incompatible with that of the selected emulation
/tmp/instance-1/output-1/host/opt/ext-toolchain/bin/../lib/gcc/mips64el-buildroot-linux-uclibc/8.4.0/../../../../mips64el-buildroot-linux-uclibc/bin/ld: failed to merge target specific data of file build/release/libmupdf.a(Dingbats.cff.o)

To fix it, use CC instead of LD for partial linking.

'nostdlib' has to be added, or it will try to use GCC libraries for
partial linking and fail with 'cannot find -lgcc_s'.

Fixes:
- http://autobuild.buildroot.net/results/156fe9ee5f6dccdc98990f6c5de5562383bc2b74

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>